+++ credentials_setup.sh
@@ -6,8 +6,8 @@
 #### [ Variables ]
 
 
-credentials="NadekoBot/src/NadekoBot/credentials.json"
-bak_credentials="NadekoBot/src/NadekoBot/credentials.json.bak"
+credentials_path="NadekoBot/src/NadekoBot/credentials.json"
+bak_credentials_path="NadekoBot/src/NadekoBot/credentials.json.bak"
 
 
 #### End of [ Variables ]
@@ -81,19 +81,19 @@ read -rp "Enter your Timezone DB API Key: " timedbapi
 echo "Timezone DB API Key: $timedbapi"
 echo -e "-------------\n"
 
-## Back up credentials by renameing it to 'credentials.json.bak'.
-if [[ -f $credentials ]]; then
+## Back up the current credentials by renameing the file to 'credentials.json.bak'.
+if [[ -f $credentials_path ]]; then
     echo "Backing up current 'credentials.json' as 'credentials.json.bak'..."
-    mv "$credentials" "$bak_credentials"
+    mv "$credentials_path" "$bak_credentials_path"
     echo "Creating new 'credentials.json'..."
 ## Create a new 'credentials.json' file.
 else
     echo "Creating 'credentials.json'..."
-    touch "$credentials"
-    sudo chmod +x "$credentials"
+    touch "$credentials_path"
+    sudo chmod +x "$credentials_path"
 fi
 
-## Add the credentials to 'credentials.json'.
+## Add the new credentials to the new 'credentials.json' file.
 echo "{
     \"Token\": \"$token\",
     \"OwnerIds\": [
@@ -108,7 +108,7 @@ echo "{
     \"TimezoneDbApiKey\": \"$timedbapi\",
     \"Db\": null,
     \"TotalShards\": 1 \
-}" > "$credentials"
+}" > "$credentials_path"
 
 echo -e "\n${_GREEN}Finished creating 'credentials.json'$_NC"
 read -rp "Press [Enter] to return the the installer menu"